+++ src/main/java/kitchenpos/eatinorders/presentation/OrderTableRestController.java
@@ -1,4 +1,4 @@
-package kitchenpos.eatinorders.ui;
+package kitchenpos.eatinorders.presentation;
 
 import kitchenpos.eatinorders.application.OrderTableService;
 import kitchenpos.eatinorders.domain.OrderTable;
