+++ src/main/java/kitchenpos/eatinorders/presentation/dto/EatInOrderAcceptResponse.java
@@ -1,18 +1,16 @@
-package kitchenpos.eatinorders.ui.dto;
+package kitchenpos.eatinorders.presentation.dto;
 
 import kitchenpos.eatinorders.tobe.domain.EatInOrder;
 import kitchenpos.eatinorders.tobe.domain.OrderTableId;
 import kitchenpos.eatinorders.tobe.domain.common.OrderId;
 import kitchenpos.eatinorders.tobe.domain.common.OrderLineItems;
 import kitchenpos.eatinorders.tobe.domain.common.OrderStatus;
-import kitchenpos.eatinorders.tobe.domain.common.OrderType;
 
 import java.time.LocalDateTime;
 
-public class EatInOrderCreateResponse {
-    private OrderId id;
+public class EatInOrderAcceptResponse {
 
-    private OrderType orderType;
+    private OrderId id;
 
     private OrderStatus status;
 
@@ -22,10 +20,9 @@ public class EatInOrderCreateResponse {
 
     private LocalDateTime orderDateTime;
 
-    public static EatInOrderCreateResponse from(EatInOrder order) {
-        return new EatInOrderCreateResponse(
+    public static EatInOrderAcceptResponse from(EatInOrder order) {
+        return new EatInOrderAcceptResponse(
                 order.getId(),
-                order.getType(),
                 order.getStatus(),
                 order.getOrderLineItems(),
                 order.getOrderTableId(),
@@ -33,9 +30,8 @@ public class EatInOrderCreateResponse {
         );
     }
 
-    public EatInOrderCreateResponse(OrderId id, OrderType orderType, OrderStatus status, OrderLineItems orderLineItems, OrderTableId orderTableId, LocalDateTime orderDateTime) {
+    public EatInOrderAcceptResponse(OrderId id, OrderStatus status, OrderLineItems orderLineItems, OrderTableId orderTableId, LocalDateTime orderDateTime) {
         this.id = id;
-        this.orderType = orderType;
         this.status = status;
         this.orderLineItems = orderLineItems;
         this.orderTableId = orderTableId;
@@ -44,10 +40,6 @@ public class EatInOrderCreateResponse {
 
     public OrderId getId() {
         return id;
-    }
-
-    public OrderType getOrderType() {
-        return orderType;
     }
 
     public OrderStatus getStatus() {
