+++ src/test/java/kitchenpos/eatinorders/application/tobe/OrderTableServiceTest.java
@@ -7,10 +7,10 @@ import kitchenpos.eatinorders.infra.InMemoryOrderRepository;
 import kitchenpos.eatinorders.infra.InMemoryOrderTableRepository;
 import kitchenpos.eatinorders.tobe.domain.*;
 import kitchenpos.eatinorders.tobe.domain.common.*;
-import kitchenpos.eatinorders.ui.dto.OrderTableChangeNumberOfGuestsResponse;
-import kitchenpos.eatinorders.ui.dto.OrderTableClearResponse;
-import kitchenpos.eatinorders.ui.dto.OrderTableCreateResponse;
-import kitchenpos.eatinorders.ui.dto.OrderTableSitResponse;
+import kitchenpos.eatinorders.presentation.dto.OrderTableChangeNumberOfGuestsResponse;
+import kitchenpos.eatinorders.presentation.dto.OrderTableClearResponse;
+import kitchenpos.eatinorders.presentation.dto.OrderTableCreateResponse;
+import kitchenpos.eatinorders.presentation.dto.OrderTableSitResponse;
 import kitchenpos.menus.tobe.domain.MenuId;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
